add landDataStatic.ts
